telegram-agent/requirements.txt
pyTelegramBotAPI==4.14.0
google-adk==0.1.0
cryptography==41.0.7
pyautogui==0.9.54
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
